fix critical FindBugs warning

+++ src/main/java/de/escidoc/admintool/view/EscidocPagedTable.java
@@ -237,7 +237,7 @@ public class EscidocPagedTable extends Table {
                 final int size = realContainer.size() - 1;
                 int pages = 0;
                 if (getPageLength() != 0) {
-                    pages = (int) Math.floor(0.0 + size / getPageLength());
+                    pages = (int) Math.floor(0.0 + (double) size / getPageLength());
                 }
                 firstIndex = pages * getPageLength();
             }
